+++ src/ingest_validation_tools/directory_validator.py
@@ -1,4 +1,5 @@
 from os import walk
+import re
 from fnmatch import fnmatch
 
 
@@ -12,7 +13,7 @@ def validate_directory(path, paths_dict, dataset_ignore_globs=[]):
     Given a directory path, and a dict representing valid path globs,
     raise DirectoryValidationErrors if there are errors.
     '''
-    required_globs, allowed_globs = _get_required_allowed(paths_dict)
+    required_patterns, allowed_patterns = _get_required_allowed(paths_dict)
     required_missing_errors, not_allowed_errors = ([], [])
     if not path.exists():
         raise DirectoryValidationErrors({
@@ -20,7 +21,7 @@ def validate_directory(path, paths_dict, dataset_ignore_globs=[]):
         })
     actual_paths = []
     for triple in walk(path):
-        (dir_path, dir_names, file_names) = triple
+        (dir_path, _dir_names, file_names) = triple
         # [1:] removes leading '/', if any.
         prefix = dir_path.replace(str(path), '')[1:]
         actual_paths += (
@@ -31,11 +32,13 @@ def validate_directory(path, paths_dict, dataset_ignore_globs=[]):
     for actual in actual_paths:
         if any(fnmatch(actual, glob) for glob in dataset_ignore_globs):
             continue
-        if not any(fnmatch(actual, glob) for glob in allowed_globs):
+        if not any(
+                re.fullmatch(pattern, actual)
+                for pattern in allowed_patterns):
             not_allowed_errors.append(actual)
-    for glob in required_globs:
-        if not any(fnmatch(actual, glob) for actual in actual_paths):
-            required_missing_errors.append(glob)
+    for pattern in required_patterns:
+        if not any(re.fullmatch(pattern, actual) for actual in actual_paths):
+            required_missing_errors.append(pattern)
 
     errors = {}
     if not_allowed_errors:
@@ -46,51 +49,22 @@ def validate_directory(path, paths_dict, dataset_ignore_globs=[]):
         raise DirectoryValidationErrors(errors)
 
 
-def _get_required_allowed(nested):
+def _get_required_allowed(dir_schema):
     '''
-    Given a nested dict, flatten it and separate the keys with
-    falsy values from those with truthy values.
+    Given a directory_schema, return a pair of regex lists:
+    Those regexes which are required, and those which are allowed.
 
-    >>> nested = {
-    ...   'a': 1,
-    ...   'b': {'c': 0}
-    ... }
-    >>> _get_required_allowed(nested)
-    (['a'], ['a', 'b/c'])
+    >>> schema = [
+    ...    {'pattern': 'this_is_required'},
+    ...    {'pattern': 'this_is_optional', 'required': False}
+    ... ]
+    >>> _get_required_allowed(schema)
+    (['this_is_required'], ['this_is_required', 'this_is_optional'])
 
     '''
-    flat = _flatten(nested)
-    required = []
-    allowed = []
-    for k, v in flat.items():
-        allowed.append(k)
-        if v:
-            required.append(k)
+    allowed = [
+        item['pattern'] for item in dir_schema]
+    required = [
+        item['pattern'] for item in dir_schema
+        if 'required' not in item or item['required']]
     return required, allowed
-
-
-def _flatten(nested, joiner='/'):
-    '''
-    Given a nested dict, flatten it to a single layer dict,
-    with keys constructed by chaining the keys of the original.
-
-    >>> nested = {
-    ...   'x.txt': 1,
-    ...   'a': {
-    ...     'y.txt': 0,
-    ...     'b': {
-    ...       'c': {},
-    ...       'd': {
-    ...         'z.txt': 1 }}}}
-    >>> _flatten(nested)
-    {'x.txt': 1, 'a/y.txt': 0, 'a/b/d/z.txt': 1}
-
-    '''
-    flat = {}
-    for key, value in nested.items():
-        if type(value) != dict:
-            flat[key] = value
-        else:
-            for k, v in _flatten(value).items():
-                flat[f'{key}{joiner}{k}'] = v
-    return flat